rename: PieceNameAndColor를 PieceDto로 변경하고 Dto 패키지로 이동
